feat(angular): update structure
