+++ packages/angular/src/ionic-module.ts
@@ -22,7 +22,7 @@ import {
 import { IonBackButtonDelegateDirective } from './directives/navigation/ion-back-button';
 import { IonRouterOutlet } from './directives/navigation/ion-router-outlet';
 import { IonTabs } from './directives/navigation/ion-tabs';
-import { NavDelegate } from './directives/navigation/nav-delegate';
+import { IonNavDelegate } from './directives/navigation/nav-delegate';
 import {
   RouterLinkDelegateDirective,
   RouterLinkWithHrefDelegateDirective,
@@ -50,7 +50,7 @@ const DECLARATIONS = [
   IonTabs,
   IonRouterOutlet,
   IonBackButtonDelegateDirective,
-  NavDelegate,
+  IonNavDelegate,
   RouterLinkDelegateDirective,
   RouterLinkWithHrefDelegateDirective,
 ];
