+++ packages/angular/standalone/src/index.ts
@@ -2,4 +2,4 @@ export { IonBackButton } from './navigation/back-button';
 export { IonModal } from './overlays/modal';
 export { IonPopover } from './overlays/popover';
 export { IonRouterOutlet } from './navigation/router-outlet';
-export { NavDelegate } from './navigation/nav-delegate';
+export { IonNav } from './navigation/nav-delegate';
